Create class, base functional class DONE!
