+++ .gitignore
@@ -1,2 +1,1 @@
-shaderClass.cpp
 *.swp
